% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orders.R
\name{doseComb_to_mat}
\alias{doseComb_to_mat}
\title{Title: Get 6 complete orderings for toxicity and efficacy for drug combinations}
\usage{
doseComb_to_mat(doseComb, type)
}
\arguments{
\item{doseComb}{Either a numeric matrix or a numeric vector or a numeric list.}

\item{type}{A string.
condition 1: detailed dose combinations are directly given in matrix -> type = "comb"
                                       dose combinations
            dose A
            dose B
condition 2: directly given the number of levels for the two doses in vector -> type = "matrix"

Step 2: get the 6 orderings}
}
\value{
outMat -> Either a list or a matrix.
Note: each array refers to the index of drug combinations orderings)
}
\description{
Step 1: convert dose combinations into matrix
}
\references{
Wages NA, Conaway MR. Specifications of a continual reassessment method design for phase I trials of combined drugs. Pharmaceutical statistics. 2013 Jul;12(4):217-24.
\doi{10.1002/pst.1575}
}
