% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_likelihood.R
\name{empiric_GammaPriorLikelihood}
\alias{empiric_GammaPriorLikelihood}
\title{empiric model with gamma prior}
\usage{
empiric_GammaPriorLikelihood(beta, beta_shape, beta_inverse_scale, x, y)
}
\arguments{
\item{beta}{parameter.}

\item{beta_shape}{A number.}

\item{beta_inverse_scale}{A number.}

\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
l -> likelihood function
}
\description{
empiric model with gamma prior
}
