% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_likelihood.R
\name{empiric_NormalPriorLikelihood}
\alias{empiric_NormalPriorLikelihood}
\title{empiric model with normal prior}
\usage{
empiric_NormalPriorLikelihood(beta, beta_mean, beta_sd, x, y)
}
\arguments{
\item{beta}{parameter.}

\item{beta_mean}{A number.}

\item{beta_sd}{A number.}

\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
l -> likelihood function
}
\description{
empiric model with normal prior
}
