\docType{data}
\name{examples_results}
\alias{examples_results}
\title{Output dataset for examples given list of inputs}
\usage{
examples_results
}
\description{
The dataset contains 1296 rows (6 scenarios, 4 different combinations of link functions and prior distributions, 
 2 sets of skeletons, 3 maximum number of patients, 3 toxicity and efficacy correlations, and 3 subset number 
 of patients) that represent each condition by 1000 simulations, along with 15 columns that contain 9 operating 
 characteristics, other 6 columns including Scenario, Model for link function and prior distribution, N for 
 maximum number of patients, Skeleton, Nphase for subset number of patients, and corr for toxicity and efficacy 
 correlation to separate each condition.
}
\keyword{datasets}