% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_likelihood.R
\name{logisticOnePara_GammaPriorLikelihood}
\alias{logisticOnePara_GammaPriorLikelihood}
\title{one-parameter logistic model with gamma prior}
\usage{
logisticOnePara_GammaPriorLikelihood(
  alpha1,
  alpha1_shape,
  alpha1_inverse_scale,
  intcpt,
  x,
  y
)
}
\arguments{
\item{alpha1}{parameter.}

\item{alpha1_shape}{A number.}

\item{alpha1_inverse_scale}{A number.}

\item{intcpt}{A number.}

\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
l -> likelihood function
}
\description{
one-parameter logistic model with gamma prior
}
