% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_likelihood.R
\name{logisticOnePara_NormalPriorLikelihood}
\alias{logisticOnePara_NormalPriorLikelihood}
\title{one-parameter logistic model with normal prior}
\usage{
logisticOnePara_NormalPriorLikelihood(
  alpha1,
  alpha1_mean,
  alpha1_sd,
  intcpt,
  x,
  y
)
}
\arguments{
\item{alpha1}{parameter.}

\item{alpha1_mean}{A number.}

\item{alpha1_sd}{A number.}

\item{intcpt}{A number.}

\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
l -> likelihood function
}
\description{
one-parameter logistic model with normal prior
}
