% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_likelihood.R
\name{logisticTwoPara_GammaPriorLikelihood}
\alias{logisticTwoPara_GammaPriorLikelihood}
\title{two-parameter logistic model with gamma prior}
\usage{
logisticTwoPara_GammaPriorLikelihood(
  alpha0,
  alpha1,
  alpha0_shape,
  alpha0_inverse_scale,
  alpha1_shape,
  alpha1_inverse_scale,
  x,
  y
)
}
\arguments{
\item{alpha0}{parameter.}

\item{alpha1}{parameter.}

\item{alpha0_shape}{A number.}

\item{alpha0_inverse_scale}{A number.}

\item{alpha1_shape}{A number.}

\item{alpha1_inverse_scale}{A number.}

\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
l -> likelihood function
}
\description{
two-parameter logistic model with gamma prior
}
