% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_likelihood.R
\name{logisticTwoPara_NormalPriorLikelihood}
\alias{logisticTwoPara_NormalPriorLikelihood}
\title{two-parameter logistic model with normal prior}
\usage{
logisticTwoPara_NormalPriorLikelihood(
  alpha0,
  alpha1,
  alpha0_mean,
  alpha0_sd,
  alpha1_mean,
  alpha1_sd,
  x,
  y
)
}
\arguments{
\item{alpha0}{parameter.}

\item{alpha1}{parameter.}

\item{alpha0_mean}{A number.}

\item{alpha0_sd}{A number.}

\item{alpha1_mean}{A number.}

\item{alpha1_sd}{A number.}

\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
l -> likelihood function
}
\description{
two-parameter logistic model with normal prior
}
