% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_likelihood.R
\name{tanh_ExpPriorLikelihood}
\alias{tanh_ExpPriorLikelihood}
\title{Title: Bayesian likelihood inference}
\usage{
tanh_ExpPriorLikelihood(beta, beta_mean, x, y)
}
\arguments{
\item{beta}{parameter.}

\item{beta_mean}{A number.}

\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
l -> likelihood function
}
\description{
Title: Bayesian likelihood inference
}
\references{
: R package dfcrm
Hyperbolic tangent model with exponential prior
}
